+++ mods/src/creative/creative.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int creative_is_restricted();
+
+#ifdef __cplusplus
+}
+#endif