+++ mods/src/init/init.c
@@ -1,7 +1,11 @@
 #include "init.h"
 
+#include <media-layer/core.h>
+
 __attribute__((constructor)) static void init() {
+    media_ensure_loaded();
     run_tests();
+    init_version();
     init_compat();
 #ifdef MCPI_SERVER_MODE
     init_server();
@@ -22,5 +26,7 @@ __attribute__((constructor)) static void init() {
     init_options();
     init_chat();
     init_home();
-    init_version();
+#if !defined(MCPI_SERVER_MODE) && !defined(MCPI_HEADLESS_MODE)
+    init_benchmark();
+#endif
 }